requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pypdf>=4.0.0
pandas>=2.2.0
watchdog>=4.0.0
structlog>=24.1.0
python-multipart>=0.0.9
pyyaml>=6.0.0
pytesseract>=0.3.10
Pillow>=10.0.0

